requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
gunicorn==20.1.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.2
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
